Super poncho cambios en UIUI

+++ PocheWebSocket/src/co/com/poncho/websocket/UserSessionHandler.java
@@ -61,7 +61,7 @@ public class UserSessionHandler {
 		sendToAllConnectedSessions(addMessage);
 	}
 
-	public void registerVote(int voto, int tipoVoto, Session session) {
+	public void registerVote(float voto, int tipoVoto, Session session) {
 		Usuario usuario = sesionesUsuarios.get(session.getId());
 		usuario.setVoto(voto);
 		usuario.setTipoVoto(tipoVoto);
